+++ Sample/Shared/Helpers/UserInfoHelper.swift
@@ -1,0 +1,13 @@
+import Foundation
+struct UserInfoHelper {
+    enum ErrorType {
+        case duplicateTracker
+    }
+
+    static func validateInput(userID: String?, idTracker: String?, token: String?) -> ErrorType? {
+        guard idTracker.isEmpty || token.isEmpty else {
+            return .duplicateTracker
+        }
+        return nil
+    }
+}